Fix some code issues
1、Fix the abnormal branch return value error of SAL_CERT_CheckCertKeyUsage function
2、Add valueLen size judgment to the macro function "PROCESS_STRING_PARAM"
3、Fix parameter passing error when calling 'CheckCertSecuritylevel' to parse CA list
4、Add length judgment to HITLS_CFG_SetAlpnProtos interface
5、Adjust the member data types of the internal structure 'HkdfLabel'
6、CcmPrepare/AdeDecrypt determines inLen>tagLen
7、Cleaning up some sensitive information

Cherry-picked from: https://gitcode.com/openHiTLS/openhitls/merge_requests/1744

+++ tls/crypt/crypt_adapt/crypt.c
@@ -52,8 +52,8 @@ HITLS_CRYPT_KdfMethod g_cryptKdfMethod = {0};
 
 typedef struct {
     uint16_t length;        /* Length of the derived key */
-    uint8_t labelLen;       /* Label length */
-    uint8_t ctxLen;         /* Length of the context information */
+    uint32_t labelLen;       /* Label length */
+    uint32_t ctxLen;         /* Length of the context information */
     const uint8_t *label;   /* Label */
     const uint8_t *ctx;     /* Context information */
     const uint8_t *labelPrefix;
@@ -901,8 +901,8 @@ int32_t SAL_CRYPT_HkdfExpandLabel(CRYPT_KeyDeriveParameters *deriveInfo, uint8_t
 
     HkdfLabel info = {0};
     info.length = (uint16_t)outLen;
-    info.labelLen = (uint8_t)deriveInfo->labelLen;
-    info.ctxLen = (uint8_t)deriveInfo->seedLen;
+    info.labelLen = deriveInfo->labelLen;
+    info.ctxLen = deriveInfo->seedLen;
     info.label = deriveInfo->label;
     info.ctx = deriveInfo->seed;
     info.labelPrefix = deriveInfo->labelPrefix;
